fixed bug with Settings screen navigation (now there is an Activity with preference fragment instead of just preference fragment)
