+++ app/src/main/AndroidManifest.xml
@@ -35,6 +35,8 @@
 
         <activity android:name=".ui.activities.LoginActivity_" />
 
+        <activity android:name=".ui.activities.TrackerPreferenceActivity" />
+
         <activity
             android:name=".ui.activities.MainActivity_"
             android:label="@string/app_name"
