+++ app/src/main/res/menu/drawer_menu.xml
@@ -18,11 +18,6 @@
             android:icon="@drawable/ic_statistics"
             android:title="@string/nav_drawer_statistics">
         </item>
-        <item
-            android:id="@+id/drawer_item_settings"
-            android:icon="@drawable/ic_settings"
-            android:title="@string/nav_drawer_settings">
-        </item>
     </group>
     <item android:title="@string/nav_drawer_subtitle_account">
         <menu>
